backend/requirements.txt
# Backend dependencies
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# HTTP / HTML
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.8.0

# ML / Retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
numpy>=1.26.0
sentence-transformers>=2.7.0

# Gemini / Google AI (new SDK — provides `from google import genai`)
google-genai>=1.0.0

# Data
openpyxl>=3.1.0
